feat: migrate and set up tailwind on root layout
