--- pages/_document.tsx
@@ -1,15 +1,0 @@
-import { Head, Html, Main, NextScript } from "next/document";
-
-function Document() {
-  return (
-    <Html lang="es">
-      <Head></Head>
-      <body>
-        <Main />
-        <NextScript />
-      </body>
-    </Html>
-  );
-}
-
-export default Document;
